+++ src/Virtlink.Utilib/Collections/Enumerables.cs
@@ -2,6 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
+using JetBrains.Annotations;
 
 namespace Virtlink.Utilib.Collections
 {
@@ -11,6 +12,28 @@ namespace Virtlink.Utilib.Collections
     public static class Enumerables
     {
         /// <summary>
+        /// Returns the enumerable as a list, either by casting it or by wrapping it in a smart list.
+        /// </summary>
+        /// <typeparam name="T">The type of elements in the enumerable.</typeparam>
+        /// <param name="enumerable">The numerable to wrap.</param>
+        /// <returns>The wrapped enumerable.</returns>
+        /// <remarks>
+        /// The smart list tries to not fully enumerate the given enumerable whenever possible,
+        /// and will ensure the enumerable is enumerated only once.
+        /// </remarks>
+        public static IReadOnlyList<T> AsList<T>(this IEnumerable<T> enumerable)
+        {
+            #region Contract
+            if (enumerable == null)
+                throw new ArgumentNullException(nameof(enumerable));
+            #endregion
+
+            if (enumerable is IReadOnlyList<T> list)
+                return list;
+            return new SmartList<T>(enumerable);
+        }
+
+        /// <summary>
         /// Attempts to get the number of elements in the enumerable,
         /// only if it can be determined efficiently.
         /// </summary>
@@ -19,7 +42,7 @@ namespace Virtlink.Utilib.Collections
         /// <returns>The number of elements;
         /// or <see langword="null"/> when it could not be determined
         /// without enumerating all elements.</returns>
-        public static int? TryGetCount<T>(IEnumerable<T> enumerable)
+        public static int? TryGetCount<T>([NoEnumeration] IEnumerable<T> enumerable)
         {
             #region Contract
             if (enumerable == null)
@@ -34,7 +57,7 @@ namespace Virtlink.Utilib.Collections
             if (roCollection != null)
                 return roCollection.Count;
 
-            return TryGetCount((IEnumerable)enumerable);
+            return TryGetCount((IEnumerable) enumerable);
         }
 
         /// <summary>
@@ -45,7 +68,7 @@ namespace Virtlink.Utilib.Collections
         /// <returns>The number of elements;
         /// or <see langword="null"/> when it could not be determined
         /// without enumerating all elements.</returns>
-        public static int? TryGetCount(IEnumerable enumerable)
+        public static int? TryGetCount([NoEnumeration] IEnumerable enumerable)
         {
             #region Contract
             if (enumerable == null)
@@ -91,6 +114,47 @@ namespace Virtlink.Utilib.Collections
                             yield return e;
                         }
                     }
+                }
+            }
+
+            return Enumerator();
+        }
+
+        /// <summary>
+        /// Zips two sequences together, like Zip, but enforces that both sequences have equal lengths.
+        /// </summary>
+        /// <typeparam name="T1">The type of elements in the first sequence.</typeparam>
+        /// <typeparam name="T2">The type of elements in the second sequence.</typeparam>
+        /// <typeparam name="TResult">The result selector.</typeparam>
+        /// <returns>The resulting sequence.</returns>
+        public static IEnumerable<TResult> ZipEqual<T1, T2, TResult>(
+            this IEnumerable<T1> first,
+            IEnumerable<T2> second,
+            Func<T1, T2, TResult> resultSelector)
+        {
+            #region Contract
+            if (first == null)
+                throw new ArgumentNullException(nameof(first));
+            if (second == null)
+                throw new ArgumentNullException(nameof(second));
+            if (resultSelector == null)
+                throw new ArgumentNullException(nameof(resultSelector));
+            #endregion
+
+            IEnumerable<TResult> Enumerator()
+            {
+                using (var e1 = first.GetEnumerator())
+                using (var e2 = second.GetEnumerator())
+                {
+                    while (e1.MoveNext())
+                    {
+                        if (!e2.MoveNext())
+                            throw new InvalidOperationException("Second sequence is shorter than the first.");
+
+                        yield return resultSelector(e1.Current, e2.Current);
+                    }
+                    if (e2.MoveNext())
+                        throw new InvalidOperationException("First sequence is shorter than the second.");
                 }
             }